requirements.txt
crewai
langchain
langchain-google-genai
python-dotenv
requests